+++ books.json
@@ -1,1 +1,1 @@
-[{"title":"Kokoko","author":"Kokokoko"},{"title":"Gygygygyg","author":"Ygygyygyggygyg"},{"title":"Huhuhuhuh","author":"Fttftftf"},{"title":"Geo","author":"Mwafrika"},{"title":"Rising sun","author":"Drie"}]
+[{"title":"Kokoko","author":"Kokokoko"},{"title":"Gygygygyg","author":"Ygygyygyggygyg"},{"title":"Huhuhuhuh","author":"Fttftftf"},{"title":"Geo","author":"Mwafrika"},{"title":"Rising sun","author":"Drie"},{"title":"W","author":"E"}]
