+++ people.json
@@ -1,1 +1,1 @@
-[{"name":"kjjijijij","id":652,"age":"8","parent_permission":true},{"name":"jiiji","id":721,"age":"87","parent_permission":true},{"name":"koko","id":362,"age":"8","parent_permission":true},{"name":"kokkokkok","id":92,"age":"76","parent_permission":true},{"name":"joe","id":537,"age":"65","parent_permission":true},{"name":"huhuh","id":347,"age":"7","parent_permission":true},{"name":"hhhh","id":876,"age":"67","parent_permission":true},{"name":"student1","id":599,"age":"45","parent_permission":true},{"name":"teacher","id":343,"age":"54","parent_permission":true},{"name":"Dire","id":338,"age":"37","parent_permission":true},{"name":"g","id":791,"age":"12","parent_permission":true}]
+[{"name":"kjjijijij","id":957,"age":"8"},{"name":"jiiji","id":985,"age":"87"},{"name":"koko","id":246,"age":"8"},{"name":"kokkokkok","id":454,"age":"76"},{"name":"joe","id":720,"age":"65"},{"name":"huhuh","id":239,"age":"7"},{"name":"hhhh","id":353,"age":"67"},{"name":"student1","id":235,"age":"45"},{"name":"teacher","id":739,"age":"54"},{"name":"Dire","id":418,"age":"37"},{"name":"g","id":15,"age":"12"},{"name":"w","id":354,"age":"12"},{"name":"e","id":733,"age":"32"}]
